fixed compiler warning in multifilereader
- fixed:
..\..\..\main\lib\src\MultiFileReader.cc(28): warning C4018: '<' : signed/unsigned mismatch [C:\Users\hperrey\Desktop\eudaq\build\main\lib\EUDAQ.vcxproj]

+++ main/lib/src/MultiFileReader.cc
@@ -25,7 +25,7 @@ bool eudaq::multiFileReader::NextEvent( size_t skip /*= 0*/ )
 	{
 	
 	do{
-		for (int fileID = 0; fileID < m_fileReaders.size(); ++fileID)
+		for (size_t fileID = 0; fileID < m_fileReaders.size(); ++fileID)
 		{
 			if (!m_fileReaders[fileID]->NextEvent()&&m_sync.SubEventQueueIsEmpty(fileID))
 			{
